# turnos/contracts/openapi.json
{
  "components" : {
    "schemas" : {
...
      "CreateGroup" : {
        "properties" : {
          "name" : {
            "maxLength" : 40,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "name" ],
        "type" : "object"
      },
      "CreateInvite" : {
        "properties" : {
          "email" : {
            "format" : "email",
            "type" : "string"
          },
          "maxUses" : {
            "format" : "int32",
            "maximum" : 100,
            "minimum" : 1,
            "type" : "integer"
          },
          "name" : {
            "maxLength" : 120,
            "minLength" : 0,
            "type" : "string"
          },
          "rank" : {
            "maxLength" : 40,
            "minLength" : 0,
            "type" : "string"
          },
          "role" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "CreatePosto" : {
        "properties" : {
          "location" : {
            "maxLength" : 80,
            "minLength" : 0,
            "type" : "string"
          },
          "name" : {
            "maxLength" : 120,
            "minLength" : 0,
            "type" : "string"
          },
          "timeZone" : {
            "type" : "string"
          }
        },
        "required" : [ "location", "name" ],
        "type" : "object"
      },
...
      "GroupView" : {
        "properties" : {
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "members" : {
            "items" : {
              "$ref" : "#/components/schemas/MemberView"
            },
            "type" : "array"
          },
          "name" : {
            "type" : "string"
          }
        },
        "required" : [ "id", "members", "name" ],
        "type" : "object"
      },
...
      "InvitePreview" : {
        "properties" : {
          "commander" : {
            "type" : "boolean"
          },
          "email" : {
            "type" : "string"
          },
          "expiresAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "groupId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "groupName" : {
            "type" : "string"
          },
          "invitedBy" : {
            "type" : "string"
          },
          "link" : {
            "type" : "boolean"
          },
          "name" : {
            "type" : "string"
          },
          "postoName" : {
            "type" : "string"
          },
          "rank" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          }
        },
        "required" : [ "commander", "email", "expiresAt", "groupId", "groupName", "invitedBy", "link", "name", "postoName", "rank", "status" ],
        "type" : "object"
      },
      "InviteView" : {
        "properties" : {
          "acceptedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "expiresAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "link" : {
            "type" : "boolean"
          },
          "maxUses" : {
            "format" : "int32",
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          },
          "rank" : {
            "type" : "string"
          },
          "role" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "uses" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "required" : [ "acceptedAt", "email", "expiresAt", "id", "link", "maxUses", "name", "rank", "role", "status", "uses" ],
        "type" : "object"
      },
...
      "MemberView" : {
        "properties" : {
          "commander" : {
            "type" : "boolean"
          },
          "displayName" : {
            "type" : "string"
          },
          "rank" : {
            "type" : "string"
          },
          "serviceNumber" : {
            "type" : "string"
          },
          "userId" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "required" : [ "commander", "displayName", "rank", "serviceNumber", "userId" ],
        "type" : "object"
      },
      "MembershipView" : {
        "properties" : {
          "commander" : {
            "type" : "boolean"
          },
          "groupId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "groupName" : {
            "type" : "string"
          },
          "postoId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "postoName" : {
            "type" : "string"
          }
        },
        "required" : [ "commander", "groupId", "groupName", "postoId", "postoName" ],
        "type" : "object"
      },
...
      "PasswordResetView" : {
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "expiresAt" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "required" : [ "code", "expiresAt" ],
        "type" : "object"
      },
...
      "PostoView" : {
        "properties" : {
          "groups" : {
            "items" : {
              "$ref" : "#/components/schemas/GroupView"
            },
            "type" : "array"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "location" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "timeZone" : {
            "type" : "string"
          }
        },
        "required" : [ "groups", "id", "location", "name", "timeZone" ],
        "type" : "object"
      },
...
      "TransferCommand" : {
        "properties" : {
          "userId" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "required" : [ "userId" ],
        "type" : "object"
      },
...
    }
  },
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/groups/{groupId}/commander" : {
      "post" : {
        "operationId" : "transfer",
        "parameters" : [ {
          "in" : "path",
          "name" : "groupId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TransferCommand"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "tags" : [ "units-controller" ]
      }
    },
    "/api/v1/groups/{groupId}/invites" : {
      "get" : {
        "operationId" : "invites",
        "parameters" : [ {
          "in" : "path",
          "name" : "groupId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/InviteView"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "units-controller" ]
      },
      "post" : {
        "operationId" : "invite",
        "parameters" : [ {
          "in" : "path",
          "name" : "groupId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateInvite"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : { },
                  "type" : "object"
                }
              }
            },
            "description" : "Created"
          }
        },
        "tags" : [ "units-controller" ]
      }
    },
    "/api/v1/groups/{groupId}/invites/{inviteId}" : {
      "delete" : {
        "operationId" : "revoke",
        "parameters" : [ {
          "in" : "path",
          "name" : "groupId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "inviteId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "tags" : [ "units-controller" ]
      }
    },
    "/api/v1/groups/{groupId}/members/{userId}" : {
      "delete" : {
        "operationId" : "removeMember",
        "parameters" : [ {
          "in" : "path",
          "name" : "groupId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "userId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "tags" : [ "units-controller" ]
      }
    },
    "/api/v1/groups/{groupId}/members/{userId}/password-reset" : {
      "post" : {
        "operationId" : "passwordReset",
        "parameters" : [ {
          "in" : "path",
          "name" : "groupId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "userId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PasswordResetView"
                }
              }
            },
            "description" : "Created"
          }
        },
        "tags" : [ "units-controller" ]
      }
    },
    "/api/v1/invites/{code}" : {
      "get" : {
        "operationId" : "preview",
        "parameters" : [ {
          "in" : "path",
          "name" : "code",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/InvitePreview"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "units-controller" ]
      }
    },
    "/api/v1/invites/{code}/accept" : {
      "post" : {
        "operationId" : "accept_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "code",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MembershipView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "units-controller" ]
      }
    },
...
    "/api/v1/postos" : {
      "get" : {
        "operationId" : "postos",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/PostoView"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "units-controller" ]
      },
      "post" : {
        "operationId" : "createPosto",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreatePosto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PostoView"
                }
              }
            },
            "description" : "Created"
          }
        },
        "tags" : [ "units-controller" ]
      }
    },
    "/api/v1/postos/{postoId}" : {
      "get" : {
        "operationId" : "posto",
        "parameters" : [ {
          "in" : "path",
          "name" : "postoId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PostoView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "units-controller" ]
      }
    },
    "/api/v1/postos/{postoId}/groups" : {
      "post" : {
        "operationId" : "createGroup",
        "parameters" : [ {
          "in" : "path",
          "name" : "postoId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateGroup"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GroupView"
                }
              }
            },
            "description" : "Created"
          }
        },
        "tags" : [ "units-controller" ]
      }
    },
...
  },
  "servers" : [ {
    "url" : "/"
  } ]
}